# docs/openapi/business-api.yaml
openapi: 3.1.0
info:
  title: InterviewCopilot Business API
  version: 1.0.0
  description: >-
    Browser-facing V1 contract owned by the Java business service. The Java
    service is the sole business system of record and the only API called by
    the frontend.
servers:
  - url: /api/v1
    description: Version 1 business API
...
security:
  - bearerAuth: []
paths:
...
  /interviews:
    post:
      tags: [Interviews]
      summary: Create an interview
      description: >-
        Creates an interview in CREATED state. Replaying the same request with
        the same idempotency key returns the original resource with HTTP 200.
      operationId: createInterview
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInterviewRequest'
      responses:
        '200':
          description: Original interview returned for an idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewResponse'
        '201':
          description: Interview created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Interviews]
      summary: List interview history
      description: Returns only interviews owned by the authenticated user.
      operationId: listInterviews
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/InterviewStatusQuery'
      responses:
        '200':
          description: Interview page returned in reverse creation order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /interviews/{interviewId}:
    parameters:
      - $ref: '#/components/parameters/InterviewIdPath'
    get:
      tags: [Interviews]
      summary: Get interview details
      description: >-
        The service must enforce ownership by matching both interviewId and
        the authenticated user. Internal question-generation data is omitted.
      operationId: getInterview
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Interview details returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /interviews/{interviewId}/start:
    parameters:
      - $ref: '#/components/parameters/InterviewIdPath'
    post:
      tags: [Interviews]
      summary: Start an interview
      description: >-
        Moves a CREATED interview to IN_PROGRESS and obtains one first
        question. A replay returns the existing question or resumes generation
        when the state changed but no question was saved.
      operationId: startInterview
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '200':
          description: Interview started or existing start result replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartInterviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /interviews/{interviewId}/current-question:
    parameters:
      - $ref: '#/components/parameters/InterviewIdPath'
    get:
      tags: [Interviews]
      summary: Get the current question
      description: >-
        Returns the current public question. A completed interview returns
        interview_status COMPLETED with a null question.
      operationId: getCurrentQuestion
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Current question state returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentQuestionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /interviews/{interviewId}/answers:
    parameters:
      - $ref: '#/components/parameters/InterviewIdPath'
    post:
      tags: [Interviews]
      summary: Submit the current answer
      description: >-
        Saves one immutable answer, evaluates it, and either obtains one next
        question or completes the interview. A replay never creates a second
        answer or evaluation.
      operationId: submitAnswer
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Original result returned for an idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitAnswerResponse'
        '201':
          description: Answer accepted and processing completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitAnswerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /interviews/{interviewId}/cancel:
    parameters:
      - $ref: '#/components/parameters/InterviewIdPath'
    post:
      tags: [Interviews]
      summary: Cancel an interview
      description: Only CREATED or IN_PROGRESS interviews can be cancelled.
      operationId: cancelInterview
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '200':
          description: Interview cancelled or cancellation result replayed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelInterviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /interviews/{interviewId}/report:
    parameters:
      - $ref: '#/components/parameters/InterviewIdPath'
    get:
      tags: [Reports]
      summary: Get an interview report
      description: >-
        Once the interview is COMPLETED this operation always returns HTTP 200
        with deterministic metrics. AI summary fields are null until the
        report is AVAILABLE. Before completion it returns HTTP 409.
      operationId: getInterviewReport
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Report state and deterministic metrics returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Java-issued access JWT with a 60-minute lifetime.
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: Optional caller request ID; the service generates one when absent.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Stable key for this user and operation. Reuse with a different request
        payload returns IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[!-~]+$'
...
  responses:
    BadRequest:
      description: Request syntax or framing is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Difficulty:
      type: string
      enum: [EASY, MEDIUM, HARD]
...
    CreateInterviewRequest:
      type: object
      additionalProperties: false
      required: [target_position, skills, difficulty]
      properties:
        target_position:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^\S(?:.*\S)?$'
        skills:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          description: Skills are trimmed and deduplicated case-insensitively.
          items:
            type: string
            minLength: 1
            maxLength: 50
            pattern: '^\S(?:.*\S)?$'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        question_count:
          type: integer
          minimum: 3
          maximum: 10
          default: 5
      example:
        target_position: Java Backend Engineer
        skills: [Java, Spring Boot, MySQL]
        difficulty: MEDIUM
        question_count: 5
...